refactor(deno-runtime): move prepareEnvironment to main.ts
